News-Spot/requirements.txt
Flask==2.2.5
requests==2.31.0
newspaper3k==0.2.8
nltk==3.8.1
scikit-learn==1.3.1
gunicorn==20.1.0
lxml[html_clean]==4.9.3

